19
19

+++ The Rainbow Dinosaur/Assets/Scripts/Game/Entity/Behavior/Condition/HpChanged.cs
@@ -3,7 +3,7 @@ namespace NBehaviour.NCondition
 {
 	public class HpChanged :Condition
 	{
-		int hpBefore;
+		float hpBefore;
 		public HpChanged (Entity entity):base(entity)
 		{
 			hpBefore = entity.hp;
